feat: integrate dashboards into unified blueprint
